Trait substrate_consensus_aura::SyncOracle

pub trait SyncOracle {
    fn is_major_syncing(&self) -> bool;

    fn is_offline(&self) -> bool;
}

An oracle for when major synchronization work is being undertaken.

Generally, consensus authoring work isn't undertaken while well behind the head of the chain.

Required methods

fn is_major_syncing(&self) -> bool

Whether the synchronization service is undergoing major sync. Returns true if so.

fn is_offline(&self) -> bool

Whether the synchronization service is offline. Returns true if so.

Loading content...